Sav 5 bytes per Animation trigger

+++ GDEdit/GDEdit/Utilities/Objects/GeometryDash/LevelObjects/Triggers/AnimateTrigger.cs
@@ -15,10 +15,10 @@ namespace GDEdit.Utilities.Objects.GeometryDash.LevelObjects.Triggers
         public override int ObjectID => (int)Enumerations.GeometryDash.TriggerType.Animate;
         
         /// <summary>The target Group ID of the trigger.</summary>
-        public int TargetGroupID { get; set; }
+        public short TargetGroupID { get; set; }
         /// <summary>The Animation ID property of the trigger.</summary>
         [ObjectStringMappable(ObjectParameter.AnimationID)]
-        public int AnimationID { get; set; }
+        public byte AnimationID { get; set; }
 
         /// <summary>Initializes a new instance of the <seealso cref="AnimateTrigger"/> class.</summary>
         public AnimateTrigger() { }
